refactor(commands): extract graph, facts, analytics, export domains (gated, verbatim)
Pure MOVE, zero behavior change. Four gate-touching command domains
lift out of commands/mod.rs into commands/{graph,facts,analytics,export}.rs
(each #[path]-aliased — the names shadow crate::{graph,facts,export});
mod.rs -1262 lines (37,210 -> 35,948), lib.rs UNTOUCHED (glob re-export).

All 40 moved #[tauri::command]s + helpers are BYTE-IDENTICAL to trunk
(verbatim bodies; the only change is fn -> pub(crate) on the few a
mod.rs-resident test/command still calls via the re-export). The gate/
mask logic is UNCHANGED — only relocated: get_full_graph / get_backlinks
/ get_entity_detail (graph), the user-memory + supersession + gated
memory-brief (facts), the audit-finding + egress-ledger + audit_row_visible
re-gate (analytics), export_audio/master/note/canvas (export). The shared
gate helpers unlocked_snapshot + assert_in_vault stay in mod.rs, promoted
pub(crate) (bodies byte-identical); meeting_is_unlocked/visibility_clause
already pub(crate). No gate helper was moved.

Tests stay in mod.rs (lock_read_gate_tests/lifecycle_tests/entity_dossier
_tests); #[test] 344 + 2 tokio on trunk and now. Verified: targeted cargo
test --lib graph/entity/user_memor/fact/analytics/audit/egress/export/
supersession/dossier/memory/lock_read_gate all green (incl. the sealed/
visible/gate tests that prove the gate still fires post-move), clippy
--lib -D warnings clean, MSRV grep clean.
