+++ src-tauri/src/commands/mod.rs
@@ -58,6 +58,39 @@ pub use mcp_commands::*;
 mod settings_commands;
 pub use settings_commands::*;
 
+// Graph / entity / people / dossier READ commands (a GATED domain — every read snapshots the live
+// `unlocked` set and pushes it through the DB visibility predicate; the gate LOGIC is byte-identical,
+// only relocated). Bound as `graph_commands` (via `#[path]`) to avoid colliding with the crate-level
+// `crate::links`/graph types — and, more importantly, so it never shadows a `graph` name.
+#[path = "graph.rs"]
+mod graph_commands;
+pub use graph_commands::*;
+
+// User-chosen-path & Canvas EXPORT commands (a GATED domain — every command fails CLOSED on
+// `meeting_is_unlocked` for a sealed-and-not-session-unlocked meeting; the gate LOGIC is
+// byte-identical, only relocated). Bound as `export_commands` (via `#[path]`) to keep it clearly
+// distinct from the crate-level `export` module. The note→VAULT export cluster + shared vault helpers
+// deliberately STAY in `commands/mod.rs`.
+#[path = "export.rs"]
+mod export_commands;
+pub use export_commands::*;
+
+// Analytics / egress-ledger / Vault-Audit commands (a GATED domain — `get_analytics` is
+// visible-content-only, the audit inbox re-gates every row against the live unlock set; the gate LOGIC
+// is byte-identical, only relocated). Bound as `analytics_commands` (via `#[path]`) to keep it clearly
+// distinct from the crate-level `audit` module and to avoid any future name shadow.
+#[path = "analytics.rs"]
+mod analytics_commands;
+pub use analytics_commands::*;
+
+// Facts / cross-meeting USER MEMORY commands + post-summary persistence (a GATED domain — user-memory
+// reads/injection filter on the live unlock snapshot; the gate LOGIC is byte-identical, only
+// relocated). Bound as `facts_commands` (via `#[path]`) to avoid colliding with the crate-level
+// `crate::facts` module imported below (E0255).
+#[path = "facts.rs"]
+mod facts_commands;
+pub use facts_commands::*;
+
 /// Keychain account for the AI Gateway API key (matches `summarize::GATEWAY_KEY_ACCOUNT`).
 /// Strictly separate from `ANTHROPIC_KEY_ACCOUNT` (defined in `commands/secrets.rs`) — never a
 /// fallback to the Anthropic key (R3). Kept here because the gateway model-listing helpers below
@@ -5354,75 +5387,6 @@ pub(crate) fn apply_organize_plan_inner(
 // because `list_user_facts_visible` filters by source-meeting `visibility_clause` on the live
 // unlocked snapshot. Forget/clear are bitemporal INVALIDATE (close valid_to), never a silent delete.
 
-/// List the current user-memory facts (open + visible) with provenance, plus the synthesized brief
-/// that is injected into grounding. GATED: only facts whose SOURCE meeting is visible under the live
-/// unlocked snapshot are returned — a sealed-not-unlocked meeting's user memory surfaces NOTHING.
-#[tauri::command]
-pub fn get_user_memory(
-    state: State<'_, AppState>,
-) -> Result<crate::user_memory::UserMemory, AppError> {
-    get_user_memory_inner(state.inner())
-}
-
-/// Inner of [`get_user_memory`] taking `&AppState` (unit-testable gate).
-pub(crate) fn get_user_memory_inner(
-    state: &AppState,
-) -> Result<crate::user_memory::UserMemory, AppError> {
-    // FLAG: memory turned OFF entirely ⇒ the explicit disabled marker (empty facts + empty brief +
-    // `disabled: true`), so the FE shows a "memory is off" affordance and NOTHING is surfaced. This
-    // mirrors the injection paths, which are also flag-suppressed — the audit view can never show
-    // facts the brain would not inject.
-    if !user_memory_enabled(state) {
-        return Ok(crate::user_memory::UserMemory::disabled());
-    }
-    let unlocked = unlocked_snapshot(state)?;
-    let facts = state.db.list_user_facts_visible(&unlocked)?;
-    // The audit view and the injected brief are derived from EXACTLY the same visible set, so the UI
-    // faithfully mirrors what the brain actually injects.
-    let brief = crate::user_memory::synthesize_brief(&facts);
-    let dtos = facts
-        .iter()
-        .map(crate::user_memory::UserMemoryFact::from_fact)
-        .collect();
-    Ok(crate::user_memory::UserMemory {
-        facts: dtos,
-        brief,
-        disabled: false,
-    })
-}
-
-/// Forget ONE user-memory fact (bitemporal invalidate — the row is CLOSED, never silently deleted,
-/// so history is preserved). After this the fact drops out of `get_user_memory` and the regenerated
-/// brief. Idempotent. Content-free logging (the fact id only, never its text).
-#[tauri::command]
-pub fn forget_user_fact(state: State<'_, AppState>, id: String) -> Result<(), AppError> {
-    forget_user_fact_inner(state.inner(), &id)
-}
-
-/// Inner of [`forget_user_fact`] taking `&AppState` (unit-testable).
-pub(crate) fn forget_user_fact_inner(state: &AppState, id: &str) -> Result<(), AppError> {
-    let at = chrono::Utc::now().to_rfc3339();
-    let closed = state.db.forget_user_fact(id, &at)?;
-    tracing::info!(target: "user_memory", fact_id = %id, closed, "user fact forgotten (invalidated)");
-    Ok(())
-}
-
-/// Clear ALL user memory: bitemporal-close every currently-open user fact (invalidate, never delete —
-/// closed history stays). After this `get_user_memory` and the brief are empty. Content-free logging
-/// (a count only).
-#[tauri::command]
-pub fn clear_user_memory(state: State<'_, AppState>) -> Result<(), AppError> {
-    clear_user_memory_inner(state.inner())
-}
-
-/// Inner of [`clear_user_memory`] taking `&AppState` (unit-testable).
-pub(crate) fn clear_user_memory_inner(state: &AppState) -> Result<(), AppError> {
-    let at = chrono::Utc::now().to_rfc3339();
-    let n = state.db.clear_user_facts(&at)?;
-    tracing::info!(target: "user_memory", count = n, "user memory cleared (all facts invalidated)");
-    Ok(())
-}
-
 /// Brain v2 L2.3 — IMPORT pasted memories from another AI assistant (a ChatGPT/Claude "what I
 /// remember about you" export) into the user-memory store. Returns the number of NEW facts added.
 ///
@@ -5735,107 +5699,6 @@ pub async fn chat_meeting(
     let (system, user) =
         crate::summarize::chat::build(&transcript, &history, &question, &memory_brief);
     provider.complete(&system, &user).await
-}
-
-/// Copy a meeting's recording (WAV) to a user-chosen path (FE picks it via a save dialog).
-#[tauri::command]
-pub fn export_audio(
-    state: State<'_, AppState>,
-    meeting_id: String,
-    dest_path: String,
-) -> Result<(), AppError> {
-    // Phase 0.5 READ-GATE: refuse to export the audio of a sealed-and-not-unlocked meeting. Its
-    // WAV is AES-GCM-encrypted at rest (audio_path → <file>.enc) and there is no plaintext on disk
-    // to copy until the folder is session-unlocked; fail closed with a Locked error.
-    if !meeting_is_unlocked(state.inner(), &meeting_id)? {
-        return Err(AppError::Locked(
-            "this meeting's folder is locked — unlock it to export the audio".into(),
-        ));
-    }
-    let meeting = state
-        .db
-        .get_meeting(&meeting_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no meeting with id {meeting_id}")))?;
-    let src = meeting
-        .audio_path
-        .ok_or_else(|| AppError::InvalidArg("this meeting has no audio file".into()))?;
-    std::fs::copy(&src, &dest_path)
-        .map_err(|e| AppError::Storage(format!("copy audio failed: {e}")))?;
-    Ok(())
-}
-
-/// Which per-stream master to export.
-enum MasterStream {
-    Mic,
-    Sys,
-}
-
-/// Shared READ-GATED export for a per-stream master archive (faithful float32 WAV). Refuses a
-/// sealed-and-not-unlocked meeting (the master is `.enc` at rest, no plaintext to copy) and never
-/// hands a path to the FE — the masters are reachable ONLY through these gated commands.
-fn export_master(
-    state: State<'_, AppState>,
-    meeting_id: &str,
-    dest_path: &str,
-    which: MasterStream,
-) -> Result<(), AppError> {
-    if !meeting_is_unlocked(state.inner(), meeting_id)? {
-        return Err(AppError::Locked(
-            "this meeting's folder is locked — unlock it to export the master".into(),
-        ));
-    }
-    let (mic, sys) = state.db.get_meeting_master_paths(meeting_id)?;
-    let src = match which {
-        MasterStream::Mic => mic,
-        MasterStream::Sys => sys,
-    }
-    .ok_or_else(|| AppError::InvalidArg("this meeting has no master for that stream".into()))?;
-    std::fs::copy(&src, dest_path)
-        .map_err(|e| AppError::Storage(format!("copy master failed: {e}")))?;
-    Ok(())
-}
-
-/// Export a meeting's MIC master archive (faithful native-rate float32 WAV) to a chosen path.
-#[tauri::command]
-pub fn export_mic_master(
-    state: State<'_, AppState>,
-    meeting_id: String,
-    dest_path: String,
-) -> Result<(), AppError> {
-    export_master(state, &meeting_id, &dest_path, MasterStream::Mic)
-}
-
-/// Export a meeting's SYSTEM master archive (faithful 48 kHz float32 WAV) to a chosen path.
-#[tauri::command]
-pub fn export_sys_master(
-    state: State<'_, AppState>,
-    meeting_id: String,
-    dest_path: String,
-) -> Result<(), AppError> {
-    export_master(state, &meeting_id, &dest_path, MasterStream::Sys)
-}
-
-/// Write a meeting's note markdown to a user-chosen path (FE picks it via a save dialog).
-#[tauri::command]
-pub fn export_note(
-    state: State<'_, AppState>,
-    meeting_id: String,
-    dest_path: String,
-) -> Result<(), AppError> {
-    // D4 READ-GATE: refuse to export a sealed-and-not-unlocked meeting's note (its plaintext
-    // markdown is blanked while sealed — exporting would write an empty/garbage file). Fail closed.
-    if !meeting_is_unlocked(state.inner(), &meeting_id)? {
-        return Err(AppError::Locked(
-            "this meeting's folder is locked — unlock it to export the note".into(),
-        ));
-    }
-    let note = state
-        .db
-        .get_latest_note_for_meeting(&meeting_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no note for meeting {meeting_id}")))?;
-    std::fs::write(&dest_path, note.markdown.as_bytes())
-        .map_err(|e| AppError::Storage(format!("write note failed: {e}")))?;
-    Ok(())
 }
 
 /// Best-effort detection of a running meeting app (Zoom / Teams / Webex) to offer a
@@ -6275,213 +6138,6 @@ pub async fn build_and_persist_entities(
     auto_link_semantic_best_effort(state, crate::links::LinkKind::Meeting, meeting_id);
 
     Ok(payload)
-}
-
-/// brain2 R2 — extract → reconcile → apply bitemporal FACTS for one summarized meeting. Pulled out
-/// of [`build_and_persist_entities`] so it can fail in isolation (its caller logs + swallows): a
-/// facts hiccup must NEVER block the note pipeline. Steps:
-///   1. BEST-EFFORT extract candidates from the note about `entity_refs` (empty with the stub/no
-///      model — the deterministic core is what carries the value),
-///   2. load the EXISTING facts for those entities (un-gated lifecycle read),
-///   3. run the PURE deterministic [`crate::facts::reconcile_facts`] at the meeting's time,
-///   4. stamp the source meeting onto the Add ops and apply them in ONE atomic tx.
-///
-/// `at` is the meeting's `started_at` (the fact's valid-time origin), falling back to now.
-fn persist_facts_for_meeting(
-    state: &AppState,
-    meeting_id: &str,
-    title: &str,
-    markdown: &str,
-    entity_refs: &[(String, String)],
-) -> Result<(), AppError> {
-    if entity_refs.is_empty() {
-        return Ok(());
-    }
-    // 1) Best-effort extraction (panic-free, empty on stub/no model/decode failure). The reasoner
-    //    is re-resolved from the LIVE config, so a consent/backend change applies without restart.
-    let candidates = crate::facts::extract_fact_candidates(
-        // Brain Live ON ⇒ the LOCAL light engine (facts stop egressing); OFF ⇒ today's Notes reasoner.
-        &*state.reasoner.extraction_reasoner(),
-        title,
-        markdown,
-        entity_refs,
-        // Post-call extraction over the full note: no tight cap (the realtime path uses a capped preset).
-        crate::reason::GenOptions::default(),
-    );
-    if candidates.is_empty() {
-        return Ok(()); // nothing to reconcile — common in the default (no-model) build.
-    }
-    // 2) Existing facts for exactly these entities.
-    let entity_ids: Vec<String> = entity_refs.iter().map(|(id, _)| id.clone()).collect();
-    let existing = state.db.facts_for_entities(&entity_ids)?;
-    // 3) Deterministic reconcile at the meeting's time (valid-time origin).
-    let at = state
-        .db
-        .get_meeting(meeting_id)?
-        .map(|m| m.started_at)
-        .unwrap_or_else(|| chrono::Utc::now().to_rfc3339());
-    let mut ops = crate::facts::reconcile_facts(&existing, &candidates, &at);
-    // 4) Stamp the source meeting (gating + purge anchor) and apply atomically.
-    crate::facts::set_meeting_id(&mut ops, meeting_id);
-    state.db.apply_fact_ops(&ops)?;
-    // Re-Truth (the vault heals itself): capture SUPERSESSIONS — every Invalidate that closed a fact
-    // sourced in an OLDER meeting is recorded for one-tap review + stamping. BEST-EFFORT and NEVER
-    // fails the note: a hiccup is logged (non-PII: counts only) and swallowed, exactly like the facts
-    // hook that wraps this whole fn.
-    if let Err(e) = record_supersessions_for_meeting(state, meeting_id, &existing, &ops) {
-        tracing::warn!(target: "retruth", error = %e, "supersession capture failed (facts unaffected)");
-    }
-    Ok(())
-}
-
-/// Re-Truth: derive + persist the supersessions for one reconcile batch (called after
-/// `apply_fact_ops`). Best-effort; returns `Ok` with nothing recorded when there are no cross-note
-/// supersessions. Non-PII log line (a count only).
-fn record_supersessions_for_meeting(
-    state: &AppState,
-    superseding_meeting_id: &str,
-    existing: &[crate::facts::Fact],
-    ops: &[crate::facts::FactOp],
-) -> Result<(), AppError> {
-    let now = chrono::Utc::now().to_rfc3339();
-    let rows = build_supersession_rows(superseding_meeting_id, existing, ops, &now);
-    if rows.is_empty() {
-        return Ok(());
-    }
-    let recorded = state.db.record_supersessions(&rows)?;
-    if recorded > 0 {
-        tracing::info!(target: "retruth", recorded, "supersessions captured for review");
-    }
-    Ok(())
-}
-
-/// PURE (no DB, no clock — `now` injected): build the SUPERSESSION rows for one reconcile batch. For
-/// each `FactOp::Invalidate` (which closed an OLD fact), resolve that old fact in `existing`
-/// (entity/predicate/old_value + its source meeting) and the matching `FactOp::Add` (the new value
-/// that superseded it, same entity+subject+predicate key reconcile used). SKIPS: an Invalidate whose
-/// old fact is absent or has no source meeting, a self-supersession (the older fact came from the SAME
-/// meeting), or a change with no matching Add. Deterministic + headless-testable.
-fn build_supersession_rows(
-    superseding_meeting_id: &str,
-    existing: &[crate::facts::Fact],
-    ops: &[crate::facts::FactOp],
-    now: &str,
-) -> Vec<crate::storage::models::SupersessionRow> {
-    use crate::facts::FactOp;
-    let norm = |s: &str| s.trim().to_lowercase();
-    let mut out = Vec::new();
-    for op in ops {
-        let FactOp::Invalidate { id, .. } = op else {
-            continue;
-        };
-        let Some(old) = existing.iter().find(|f| &f.id == id) else {
-            continue;
-        };
-        let Some(source) = old.meeting_id.as_deref().filter(|m| !m.is_empty()) else {
-            continue; // legacy/unattributed old fact — no source note to stamp.
-        };
-        if source == superseding_meeting_id {
-            continue; // a meeting refining its own earlier fact is not a cross-note supersession.
-        }
-        // The new value is the Add op sharing the old fact's (entity_id, subject, predicate) key.
-        let Some(new_value) = ops.iter().find_map(|o| match o {
-            FactOp::Add(nf)
-                if nf.entity_id == old.entity_id
-                    && norm(&nf.subject) == norm(&old.subject)
-                    && norm(&nf.predicate) == norm(&old.predicate) =>
-            {
-                Some(nf.object.clone())
-            }
-            _ => None,
-        }) else {
-            continue;
-        };
-        out.push(crate::storage::models::SupersessionRow {
-            id: uuid::Uuid::new_v4().to_string(),
-            superseding_meeting_id: superseding_meeting_id.to_string(),
-            source_meeting_id: source.to_string(),
-            entity: old.subject.clone(),
-            predicate: old.predicate.clone(),
-            old_value: old.object.clone(),
-            new_value,
-            created_at: now.to_string(),
-            applied_at: None,
-            source_pre_image: None,
-            superseding_pre_image: None,
-        });
-    }
-    out
-}
-
-/// Phase 3 CROSS-MEETING USER MEMORY — extract → reconcile → apply USER-SCOPED facts for one
-/// summarized meeting. Mirrors [`persist_facts_for_meeting`] but for the user, not entities:
-///   1. BEST-EFFORT extract user·predicate·object candidates from the note + the user's own typed
-///      notes (empty with the stub / no model — the deterministic core is what carries the value),
-///   2. load the EXISTING user facts (un-gated lifecycle read — reconcile runs before any seal),
-///   3. run the PURE deterministic [`crate::facts::reconcile_facts`] at the meeting's time (the
-///      user-scope sentinel in `entity_id` keys the reconcile),
-///   4. stamp the source meeting onto the Add ops and apply them to `user_facts` in ONE atomic tx.
-///
-/// The (derived) memory brief is regenerated lazily on the next read/turn — no cache to invalidate.
-fn persist_user_facts_for_meeting(
-    state: &AppState,
-    meeting_id: &str,
-    title: &str,
-    markdown: &str,
-) -> Result<(), AppError> {
-    // FLAG: when the user has turned cross-meeting memory OFF, skip extraction ENTIRELY — no
-    // reasoner call, no candidates, nothing new persisted. (Existing facts stay; the user can
-    // forget/clear them, and the gated reads/injection are separately flag-suppressed.)
-    if !user_memory_enabled(state) {
-        return Ok(());
-    }
-    // The user's OWN typed notes for this meeting are a high-signal memory source (an explicit
-    // "remember that…"). Empty when none (best-effort read).
-    let typed_notes = state.db.get_manual_notes(meeting_id).unwrap_or_default();
-    // D5 — the meeting's own @brain THREAD TURNS are the HIGHEST-signal source (an explicit
-    // "zapamiętaj, że…" in a thread). GATED like every content read: the just-finished meeting is
-    // its own unlocked meeting, so `list_assistant_interactions_visible` under the live unlock
-    // snapshot returns its turns (and NOTHING for a sealed-not-unlocked meeting — fail-closed). We
-    // feed the USER COMMAND text (the high-signal part), never the assistant's answer.
-    let thread_turns = gated_meeting_thread_turns(state, meeting_id);
-    // 1) Best-effort extraction (panic-free, empty on stub/no model/decode failure). The reasoner is
-    //    re-resolved from the LIVE config so a consent/backend change applies without restart.
-    let candidates = crate::user_memory::extract_user_fact_candidates(
-        // Brain Live ON ⇒ the LOCAL light engine (user facts stop egressing); OFF ⇒ today's reasoner.
-        &*state.reasoner.extraction_reasoner(),
-        title,
-        markdown,
-        &typed_notes,
-        &thread_turns,
-    );
-    if candidates.is_empty() {
-        return Ok(()); // nothing to reconcile — common in the default (no-model) build.
-    }
-    // 2) Existing user facts (all of them — the reconcile input).
-    let existing = state.db.user_facts_all()?;
-    // 3) Deterministic reconcile at the meeting's time (valid-time origin).
-    let at = state
-        .db
-        .get_meeting(meeting_id)?
-        .map(|m| m.started_at)
-        .unwrap_or_else(|| chrono::Utc::now().to_rfc3339());
-    let mut ops = crate::facts::reconcile_facts(&existing, &candidates, &at);
-    // 4) Stamp the source meeting (gating + purge anchor) and apply atomically.
-    crate::facts::set_meeting_id(&mut ops, meeting_id);
-    state.db.apply_user_fact_ops(&ops)?;
-    Ok(())
-}
-
-/// Whether cross-meeting USER MEMORY is enabled (config `user_memory_enabled`, default TRUE). When
-/// OFF: no extraction runs, no brief is injected into ANY surface, and `get_user_memory` reports the
-/// disabled marker. Fail-safe: a poisoned config mutex reports ENABLED (the default) so a transient
-/// lock error never silently disables the feature.
-fn user_memory_enabled(state: &AppState) -> bool {
-    state
-        .config
-        .lock()
-        .map(|c| c.user_memory_enabled)
-        .unwrap_or(true)
 }
 
 // ── Re-Truth (the vault heals itself) — supersession review + one-tap stamp ──────────────────────
@@ -6972,475 +6628,6 @@ pub async fn run_vault_audit(
     Ok(summary)
 }
 
-/// List findings by status (default `pending`). DEFENSIVELY re-filters each row's SOURCE
-/// visibility against the LIVE session unlock set before returning — belt-and-braces on top of
-/// purge-on-seal (which already drops pending rows, by source AND target id, inside every seal
-/// tx), so a finding whose folder sealed between pass and list can never surface.
-#[tauri::command]
-pub fn list_audit_findings(
-    state: State<'_, AppState>,
-    status: Option<String>,
-) -> Result<Vec<crate::audit::AuditFinding>, AppError> {
-    list_audit_findings_inner(state.inner(), status.as_deref())
-}
-
-pub(crate) fn list_audit_findings_inner(
-    state: &AppState,
-    status: Option<&str>,
-) -> Result<Vec<crate::audit::AuditFinding>, AppError> {
-    let status = status.unwrap_or("pending");
-    if !matches!(status, "pending" | "accepted" | "dismissed") {
-        return Err(AppError::InvalidArg(
-            "status must be pending, accepted or dismissed".into(),
-        ));
-    }
-    let unlocked = unlocked_snapshot(state)?;
-    let rows = state.db.list_audit_finding_rows(status)?;
-    let mut out = Vec::new();
-    for r in rows {
-        // SOURCE + TARGET re-gate (lock review, defense in depth): a pending row's titles/
-        // evidence reference both sides, so either side sealing between pass and list hides the
-        // row. `audit_row_visible` fails CLOSED on an untyped target (`target_id` without
-        // `target_kind` — such a row cannot be re-gated against the right table), and the
-        // explain command shares the same helper so the two gates can never diverge.
-        if !audit_row_visible(state, &r, &unlocked)? {
-            continue; // sealed since the pass — hide, never mask.
-        }
-        out.push(r.into_dto());
-    }
-    Ok(out)
-}
-
-/// Resolve one PENDING finding: `"accept"` applies its append-only vault action (re-gated at
-/// apply time), `"dismiss"` just flips the status. BOTH paths blank `evidence_md` +
-/// `accept_action` — only pending rows ever hold derived plaintext. Returns the updated finding.
-#[tauri::command]
-pub fn resolve_audit_finding(
-    app: AppHandle,
-    state: State<'_, AppState>,
-    id: String,
-    action: String,
-) -> Result<crate::audit::AuditFinding, AppError> {
-    let out = resolve_audit_finding_inner(state.inner(), &id, &action)?;
-    let pending = state.db.count_pending_audit_findings().unwrap_or(0);
-    crate::events::emit_audit_updated(&app, pending as u32);
-    Ok(out)
-}
-
-pub(crate) fn resolve_audit_finding_inner(
-    state: &AppState,
-    id: &str,
-    action: &str,
-) -> Result<crate::audit::AuditFinding, AppError> {
-    // Seal-vs-write TOCTOU: the lifecycle guard is taken INSIDE the accept path, per source kind
-    // (see `apply_audit_accept`) — NOT here. The note-canonical writer delegates to
-    // `update_note_doc_inner`, which takes the same NON-REENTRANT guard itself; holding it across
-    // that call would self-deadlock. The dismiss path and the status flip are single atomic DB
-    // statements and need no guard.
-    let row = state
-        .db
-        .get_audit_finding(id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no audit finding {id}")))?;
-    if row.status != "pending" {
-        return Err(AppError::InvalidArg("this finding was already handled".into()));
-    }
-    let status = match action {
-        "dismiss" => "dismissed",
-        "accept" => {
-            if row.accept_action.is_empty() {
-                return Err(AppError::InvalidArg(
-                    "this finding is dismiss-only — it has no accept action".into(),
-                ));
-            }
-            apply_audit_accept(state, &row)?;
-            "accepted"
-        }
-        _ => {
-            return Err(AppError::InvalidArg(
-                "action must be \"accept\" or \"dismiss\"".into(),
-            ))
-        }
-    };
-    state
-        .db
-        .resolve_audit_finding_row(id, status, chrono::Utc::now().timestamp_millis())?;
-    let updated = state
-        .db
-        .get_audit_finding(id)?
-        .ok_or_else(|| AppError::Storage("finding vanished during resolve".into()))?;
-    tracing::info!(target: "audit", finding_id = %id, kind = %updated.kind, status = %status, "audit finding resolved");
-    Ok(updated.into_dto())
-}
-
-/// APPLY one finding's accept action: an APPEND-ONLY stamp under `## Audit` in the source's
-/// exported `.md` (both sources for a contradiction). Every path RE-GATES at apply time
-/// (the prune↔seal TOCTOU discipline — a source sealed since the pass refuses with
-/// `AppError::Locked`, never stamps), and every appended `[[link]]` is re-resolved against the
-/// LIVE vault/session first (the `list_vault_titles` anti-hallucination rule).
-fn apply_audit_accept(
-    state: &AppState,
-    row: &crate::audit::AuditFindingRow,
-) -> Result<(), AppError> {
-    // Seal-vs-write TOCTOU, split by source kind: MEETING flows stamp the exported FILE and have
-    // no inner guard — hold the lifecycle guard across their gate+append (the
-    // `apply_supersessions_inner` discipline). NOTE flows write through the canonical store via
-    // `update_note_doc_inner`, which takes the SAME non-reentrant guard itself — holding it here
-    // would self-deadlock; their gate+write atomicity lives inside that writer, and a seal racing
-    // the pre-read fails CLOSED at its re-gate.
-    let _lifecycle = (row.source_kind == "meeting").then(|| lifecycle_guard(state));
-    // Re-gate the SOURCE up front, so a sealed-since-the-pass source refuses with the honest
-    // `Locked` BEFORE any per-kind validation (link re-resolution etc.) can mask it with a
-    // misleading InvalidArg. The write paths re-gate again at their own boundary.
-    let source_open = match row.source_kind.as_str() {
-        "meeting" => source_is_stampable(state, &row.source_id)?,
-        _ => {
-            let unlocked = unlocked_snapshot(state)?;
-            state.db.note_is_visible(&row.source_id, &unlocked)?
-        }
-    };
-    if !source_open {
-        return Err(AppError::Locked(
-            "this note's folder is locked — unlock it to apply the audit action".into(),
-        ));
-    }
-    let date = chrono::Utc::now().format("%Y-%m-%d").to_string();
-    match row.kind.as_str() {
-        "broken_link" => {
-            let target = row.target_title.as_deref().unwrap_or("").trim().to_string();
-            if target.is_empty() {
-                return Err(AppError::InvalidArg("this finding has no link target".into()));
-            }
-            // LIVE re-resolve (lock review): the target may have been created since the pass —
-            // the stamp would then assert a falsehood. REFUSE, loss-safe: nothing is written and
-            // the row stays pending with its evidence, so the user dismisses or re-runs the
-            // audit (chosen over auto-dismiss so no status flips without an explicit choice).
-            if broken_link_target_resolves(state, &target)? {
-                return Err(AppError::InvalidArg(
-                    "the link target now resolves — dismiss this finding or re-run the audit"
-                        .into(),
-                ));
-            }
-            // The [[..]] rides in a CODE SPAN so the stamp itself never becomes another
-            // (broken) wikilink — the one deliberate exception to "appended links resolve".
-            // Worded as an observation at audit time, not a standing claim.
-            let body = format!("> `[[{target}]]` — link target was not found at audit time");
-            stamp_audit_source(state, row, |md| {
-                crate::export::obsidian::append_audit_callout(md, &date, "broken-link", &body)
-            })
-        }
-        "stale" => {
-            // Counts recomputed LIVE through the gated reader (they may have moved since the
-            // pass); the callout carries counts only — never superseding titles (leak-safe even
-            // if a superseding source sealed since the pass).
-            let unlocked = unlocked_snapshot(state)?;
-            let facts = state
-                .db
-                .fact_rows_for_meeting_visible(&row.source_id, &unlocked)?;
-            let total = facts.len();
-            let closed = facts.iter().filter(|f| f.valid_to.is_some()).count();
-            let body = format!(
-                "> {closed} of {total} facts recorded in this note have since been superseded by later meetings"
-            );
-            stamp_audit_source(state, row, |md| {
-                crate::export::obsidian::append_audit_callout(md, &date, "stale", &body)
-            })
-        }
-        "contradiction" => {
-            let other_id = row
-                .target_id
-                .as_deref()
-                .ok_or_else(|| AppError::InvalidArg("this finding has no counterpart".into()))?;
-            // Re-gate BOTH sides BEFORE stamping either (never stamp into, or reference, a
-            // sealed note — and never leave a half-stamped pair). Both sides are MEETINGS by
-            // construction (the contradiction pass is facts-over-meetings only).
-            let source_path = audit_meeting_source_file(state, &row.source_id)?;
-            let other_path = audit_meeting_source_file(state, other_id)?;
-            let source_stem = file_stem_of(&source_path);
-            let other_stem = file_stem_of(&other_path);
-            append_to_audit_meeting_file(state, &row.source_id, &source_path, |md| {
-                crate::export::obsidian::append_audit_callout(
-                    md,
-                    &date,
-                    "conflict",
-                    &format!("> conflicting facts recorded here and in [[{other_stem}]] — review which is current"),
-                )
-            })?;
-            append_to_audit_meeting_file(state, other_id, &other_path, |md| {
-                crate::export::obsidian::append_audit_callout(
-                    md,
-                    &date,
-                    "conflict",
-                    &format!("> conflicting facts recorded here and in [[{source_stem}]] — review which is current"),
-                )
-            })
-        }
-        "unlinked_mention" => {
-            let target = row.target_title.as_deref().unwrap_or("").trim().to_string();
-            if target.is_empty() {
-                return Err(AppError::InvalidArg("this finding has no link target".into()));
-            }
-            if !audit_link_target_ok(state, &target)? {
-                return Err(AppError::InvalidArg(
-                    "the suggested link target no longer exists — dismiss this finding instead".into(),
-                ));
-            }
-            let line = format!("- Suggested link: [[{target}]]");
-            stamp_audit_source(state, row, |md| {
-                crate::export::obsidian::append_audit_line(md, &line)
-            })
-        }
-        "orphan" => {
-            // The suggestions live as [[links]] inside the (still-pending) evidence — re-extract
-            // and RE-RESOLVE each against the live vault/session; only survivors are appended.
-            let mut suggestions = Vec::new();
-            for t in crate::storage::db::extract_wikilink_titles(&row.evidence_md) {
-                if audit_link_target_ok(state, &t)? {
-                    suggestions.push(format!("[[{t}]]"));
-                }
-            }
-            if suggestions.is_empty() {
-                return Err(AppError::InvalidArg(
-                    "no suggested link still resolves — dismiss this finding instead".into(),
-                ));
-            }
-            let line = format!("- Suggested links: {}", suggestions.join(" · "));
-            stamp_audit_source(state, row, |md| {
-                crate::export::obsidian::append_audit_line(md, &line)
-            })
-        }
-        _ => Err(AppError::InvalidArg(format!(
-            "unknown finding kind {}",
-            row.kind
-        ))),
-    }
-}
-
-/// Resolve + stamp a finding's (single) source in one step, dispatched by SOURCE KIND: a
-/// meeting note is stamped on its exported `.md` (the Re-Truth precedent — the file IS its
-/// projection surface); an authored NOTE is stamped through its CANONICAL DB text (live-bug fix
-/// 2026-07-16: a Brain note created in-app may have NO exported file, and the DB text is
-/// canonical for every note regardless).
-fn stamp_audit_source(
-    state: &AppState,
-    row: &crate::audit::AuditFindingRow,
-    stamp: impl Fn(&str) -> String,
-) -> Result<(), AppError> {
-    match row.source_kind.as_str() {
-        "meeting" => {
-            let path = audit_meeting_source_file(state, &row.source_id)?;
-            append_to_audit_meeting_file(state, &row.source_id, &path, stamp)
-        }
-        "note" => stamp_audit_note_source(state, &row.source_id, stamp),
-        other => Err(AppError::InvalidArg(format!(
-            "unknown finding source kind {other}"
-        ))),
-    }
-}
-
-/// Stamp a NOTE-source finding through the note's CANONICAL store — never through the exported
-/// file. Reads the markdown through the GATED reader (sealed-and-not-session-unlocked → Locked),
-/// applies the idempotent append to the TEXT (the marker lives in the DB text), and persists via
-/// [`update_note_doc_inner`] — the note editor's own save path — so the seal-on-write gate, the
-/// re-index, and the vault re-export all ride along. The re-export CREATES the `.md` and stamps
-/// `exported_path`/`exported_hash` for a never-exported note (healing it), and `write_note`
-/// never clobbers different bytes, so an externally-edited exported file survives untouched.
-fn stamp_audit_note_source(
-    state: &AppState,
-    note_id: &str,
-    stamp: impl Fn(&str) -> String,
-) -> Result<(), AppError> {
-    let unlocked = unlocked_snapshot(state)?;
-    let Some(markdown) = state.db.note_markdown_if_visible(note_id, &unlocked)? else {
-        return Err(AppError::Locked(
-            "this note's folder is locked — unlock it to apply the audit action".into(),
-        ));
-    };
-    let written = stamp(&markdown);
-    if written == markdown {
-        return Ok(()); // already stamped — idempotent no-op.
-    }
-    let row = state
-        .db
-        .get_note_row(note_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no note {note_id}")))?;
-    let title = note_display_title(&row);
-    update_note_doc_inner(state, note_id, &title, &written)?;
-    Ok(())
-}
-
-/// Read-modify-write APPEND of one audit stamp on a MEETING note's exported `.md`: read the
-/// CURRENT file, apply the (idempotent) append, write only on change, then refresh the
-/// export-collision baseline with the Phase-0 CONDITIONAL rule
-/// ([`refresh_meeting_note_exported_hash`] — an externally-edited file keeps its stale baseline,
-/// so the next full overwrite preserves edit + stamp as a sibling instead of laundering).
-fn append_to_audit_meeting_file(
-    state: &AppState,
-    meeting_id: &str,
-    path: &str,
-    stamp: impl Fn(&str) -> String,
-) -> Result<(), AppError> {
-    let current = std::fs::read_to_string(path)
-        .map_err(|e| AppError::Export(format!("read note before audit stamp failed: {e}")))?;
-    let written = stamp(&current);
-    if written == current {
-        return Ok(()); // already stamped — idempotent no-op.
-    }
-    crate::export::obsidian::overwrite_note(std::path::Path::new(path), &written)?;
-    refresh_meeting_note_exported_hash(state, meeting_id, &current, &written)
-}
-
-/// Re-gate + resolve a MEETING finding source's exported `.md` at APPLY time (the TOCTOU
-/// re-check — the Re-Truth `source_is_stampable` posture): session-unlocked AND open-on-disk,
-/// with an exported file. Refusals are `AppError::Locked`; a missing file says what to do.
-fn audit_meeting_source_file(state: &AppState, source_id: &str) -> Result<String, AppError> {
-    if !source_is_stampable(state, source_id)? {
-        return Err(AppError::Locked(
-            "this note's folder is locked — unlock it to apply the audit action".into(),
-        ));
-    }
-    match note_file_for(state, source_id)? {
-        Some((path, _)) => Ok(path),
-        None => Err(AppError::InvalidArg(
-            "this meeting's note has no exported vault file — configure a vault and re-export the note first".into(),
-        )),
-    }
-}
-
-/// The anti-hallucination re-check for an appended `[[link]]`: the title must resolve NOW —
-/// through the GATED `resolve_wikilink` (notes/meetings/org, live session set) to a target whose
-/// folder is also open ON DISK (a link into a merely session-unlocked folder would break at
-/// relock — the `superseding_link_stem` bar), or be an existing vault file stem.
-fn audit_link_target_ok(state: &AppState, title: &str) -> Result<bool, AppError> {
-    let unlocked = unlocked_snapshot(state)?;
-    match state.db.resolve_wikilink(title, &unlocked)? {
-        Some(t) if t.kind == "meeting" => Ok(!folder_locked_on_disk(state, &t.id)?),
-        Some(t) if t.kind == "note" => {
-            let Some(row) = state.db.get_note_row(&t.id)? else {
-                return Ok(false);
-            };
-            Ok(state
-                .db
-                .folder_by_id(&row.folder_id)?
-                .map(|f| !f.locked)
-                .unwrap_or(false))
-        }
-        Some(_) => Ok(true), // org item — deliberately-disclosed, outside the folder-lock domain.
-        None => {
-            // Not a note/meeting/org — an existing user-authored vault file still resolves.
-            let Some(vault) = vault_path(state) else {
-                return Ok(false);
-            };
-            Ok(
-                crate::export::obsidian::list_vault_titles(std::path::Path::new(&vault))?
-                    .iter()
-                    .any(|t| t == title),
-            )
-        }
-    }
-}
-
-/// Ping the FE audit inbox after a purge-bearing lock/relock/delete/discard/move-into-locked
-/// mutation (count-only payload, adversarial B): pending findings vanish reactively instead of
-/// on the next manual refetch. Mirrors the screen-share relock's emit posture — best-effort by
-/// construction (`emit_audit_updated` swallows failures).
-fn emit_audit_updated_after_purge(app: &AppHandle, state: &AppState) {
-    let pending = state.db.count_pending_audit_findings().unwrap_or(0);
-    crate::events::emit_audit_updated(app, pending as u32);
-}
-
-/// Re-gate ONE finding row against a session unlock set: the SOURCE must be visible, and when a
-/// target id rides on the row its TARGET must be visible too, checked against the table its
-/// `target_kind` names. FAIL-CLOSED on an UNTYPED target (`target_id` set, `target_kind` NULL):
-/// such a row cannot be target-re-gated, so it is treated as not visible (lock-review NIT,
-/// 2026-07-16 — the previous inline check silently SKIPPED the target re-gate for untyped rows,
-/// which fails open exactly when the target seals between pass and read). Shared by the list
-/// re-gate and `explain_audit_finding` so the two can never diverge.
-fn audit_row_visible(
-    state: &AppState,
-    row: &crate::audit::AuditFindingRow,
-    unlocked: &std::collections::HashSet<String>,
-) -> Result<bool, AppError> {
-    let source_visible = match row.source_kind.as_str() {
-        "meeting" => state.db.meeting_is_visible(&row.source_id, unlocked)?,
-        _ => state.db.note_is_visible(&row.source_id, unlocked)?,
-    };
-    if !source_visible {
-        return Ok(false);
-    }
-    if let Some(tid) = &row.target_id {
-        let Some(tkind) = row.target_kind.as_deref() else {
-            return Ok(false); // untyped target — cannot be re-gated ⇒ fail closed.
-        };
-        let target_visible = match tkind {
-            "meeting" => state.db.meeting_is_visible(tid, unlocked)?,
-            _ => state.db.note_is_visible(tid, unlocked)?,
-        };
-        if !target_visible {
-            return Ok(false);
-        }
-    }
-    Ok(true)
-}
-
-// ── Vault Audit Phase 3 — weekly schedule + cloud explain ───────────────────────────────────────
-
-/// The weekly-audit schedule for the FE settings surface: the enabled flag + the last SCHEDULED
-/// run + the derived next due time (see [`crate::audit::AuditSchedule`] for the `nextDueAt`
-/// semantics). Read-only; content-free.
-#[tauri::command]
-pub fn get_audit_schedule(
-    state: State<'_, AppState>,
-) -> Result<crate::audit::AuditSchedule, AppError> {
-    get_audit_schedule_inner(state.inner())
-}
-
-pub(crate) fn get_audit_schedule_inner(
-    state: &AppState,
-) -> Result<crate::audit::AuditSchedule, AppError> {
-    let enabled = state
-        .config
-        .lock()
-        .map_err(|_| AppError::Config("config mutex poisoned".into()))?
-        .vault_audit_weekly_enabled;
-    let last_run_at = state.db.last_scheduled_audit_run_finished_at()?;
-    let next_due_at = if enabled {
-        last_run_at.map(|l| l + crate::audit::WEEKLY_AUDIT_INTERVAL_MS)
-    } else {
-        None
-    };
-    Ok(crate::audit::AuditSchedule {
-        enabled,
-        last_run_at,
-        next_due_at,
-    })
-}
-
-/// Enable/disable the weekly scheduled audit. The ONLY mutator of the flag (preserve-only on the
-/// settings DTO), persisted through the dedicated `AppConfig::set_vault_audit_weekly` (persist
-/// first, flip in-memory on durable success). Returns the updated schedule.
-#[tauri::command]
-pub fn set_audit_schedule(
-    state: State<'_, AppState>,
-    enabled: bool,
-) -> Result<crate::audit::AuditSchedule, AppError> {
-    set_audit_schedule_inner(state.inner(), enabled)
-}
-
-pub(crate) fn set_audit_schedule_inner(
-    state: &AppState,
-    enabled: bool,
-) -> Result<crate::audit::AuditSchedule, AppError> {
-    {
-        let mut cache = state
-            .config
-            .lock()
-            .map_err(|_| AppError::Config("config mutex poisoned".into()))?;
-        cache.set_vault_audit_weekly(&state.db, enabled)?;
-    }
-    tracing::info!(target: "audit", enabled, "weekly audit schedule updated");
-    get_audit_schedule_inner(state)
-}
-
 /// USER-INITIATED cloud explanation of one PENDING finding. Gates, in order:
 /// 1. the finding exists and is still pending (`InvalidArg` otherwise);
 /// 2. its source AND typed target are visible under the CURRENT session unlock set —
@@ -7581,28 +6768,6 @@ fn gated_meeting_thread_turns(state: &AppState, meeting_id: &str) -> String {
         .join("\n")
 }
 
-/// The gated cross-meeting USER MEMORY brief for injection into the non-agentic Ask / meeting-chat
-/// surfaces (design spec: parity with the @brain agentic loop, which already injects it). It is
-/// DERIVED data — never sealed, always REGENERATED from the currently-VISIBLE user facts under the
-/// passed `unlocked` snapshot — so a sealed-not-unlocked meeting's user facts inject NOTHING. When
-/// memory is disabled (config `user_memory_enabled == false`) it returns EMPTY, so the prompt is
-/// byte-identical to the pre-memory prompt. Rides the EXISTING redaction + consent egress of the
-/// surface it is injected into — no new egress class.
-///
-/// Brain v2 L2.2: `query` is the user's question when the surface has one in hand — the brief is
-/// then RELEVANCE-FILTERED (BM25 top-k over the SAME visible set, `build_memory_brief`); an empty
-/// query or zero hits falls back to the full-list brief (behavior-preserving).
-fn gated_memory_brief_for_injection(
-    state: &AppState,
-    unlocked: &std::collections::HashSet<String>,
-    query: &str,
-) -> String {
-    if !user_memory_enabled(state) {
-        return String::new();
-    }
-    crate::user_memory::build_memory_brief(&state.db, query, unlocked)
-}
-
 /// Resolve the people + projects in a meeting note → persist them to the encrypted DB graph
 /// (always) and mirror them as `[[Person]]` / `[[Project]]` vault stubs (only when a vault is
 /// configured + the meeting's folder is unsealed). The graph self-assembles. The DB sink works
@@ -7633,110 +6798,6 @@ pub async fn link_meeting_entities(
         .clone()
         .unwrap_or_else(|| "Meeting".to_string());
     build_and_persist_entities(&app, &state, &meeting_id, &title, &note.markdown).await
-}
-
-/// Max co-occurring neighbors returned with an entity's detail (the neighborhood satellites).
-const ENTITY_NEIGHBOR_LIMIT: i64 = 12;
-
-/// The self-assembling graph: all VISIBLE entity nodes (with their visible mention counts) + all
-/// VISIBLE co-occurrence edges. Snapshots the live session `unlocked` set (same as `list_folders`)
-/// and pushes it through the visibility predicate, so sealed-and-not-unlocked meetings contribute
-/// nothing — the graph can never disagree with Library/MCP about what's visible.
-#[tauri::command]
-pub fn get_graph(state: State<'_, AppState>) -> Result<GraphData, AppError> {
-    let unlocked = unlocked_snapshot(state.inner())?;
-    state.db.build_graph(&unlocked)
-}
-
-/// The FULL-BRAIN graph (Brain v3 PR-4): a SEPARATE, additive read that unifies entities + VISIBLE
-/// meetings + notes + documents as TYPED nodes and every relation (entity co-occurrence + entity→
-/// meeting mentions + `links` rows) as TYPED edges. Snapshots the live session `unlocked` set exactly
-/// like `get_graph`, so a sealed-and-not-session-unlocked meeting/note/document contributes NOTHING —
-/// no node, and no edge touching it (BOTH endpoints are gated against the visible-node set). `opts`
-/// defaults to all-off; `includeSuggested` admits un-accepted (`status='suggested'`) semantic links.
-/// Pure read: no writes, no new storage. `get_graph` is untouched + byte-compatible for its FE.
-#[tauri::command]
-pub fn get_full_graph(
-    state: State<'_, AppState>,
-    opts: Option<FullGraphOpts>,
-) -> Result<FullGraphData, AppError> {
-    let unlocked = unlocked_snapshot(state.inner())?;
-    state.db.build_full_graph(&unlocked, opts.unwrap_or_default())
-}
-
-/// `/people` personal CRM: one card per VISIBLE Person entity, rolled up over the SAME gated
-/// graph/facts/commitment readers as the graph + rollup views (`list_entities_visible` filtered to
-/// people, `entity_mentions_visible`, `list_facts_visible`, `list_open_commitments`). Snapshots the
-/// live session `unlocked` set like `get_graph`, so a person whose only mentions are in
-/// sealed-and-not-session-unlocked meetings never appears and every count reflects visible sources
-/// only. Read-only, no model, no new egress.
-#[tauri::command]
-pub fn list_people(state: State<'_, AppState>) -> Result<PeopleList, AppError> {
-    let unlocked = unlocked_snapshot(state.inner())?;
-    state.db.list_people(&unlocked)
-}
-
-/// Detail for one entity: the entity, its VISIBLE backlinked meetings (as `VaultSource` chips),
-/// and its top co-occurring neighbors. Snapshots the live `unlocked` set like `get_graph`.
-/// Errors with `InvalidArg` if the entity id is unknown.
-#[tauri::command]
-pub fn get_entity_detail(
-    state: State<'_, AppState>,
-    entity_id: String,
-) -> Result<EntityDetail, AppError> {
-    let unlocked = unlocked_snapshot(state.inner())?;
-    state
-        .db
-        .build_entity_detail(&entity_id, &unlocked, ENTITY_NEIGHBOR_LIMIT)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no entity with id {entity_id}")))
-}
-
-/// Brain v3 PR-6 — the KNOWLEDGE DIFF / decision ledger for one entity: what the vault knew AS OF
-/// `from` vs AS OF `to` (added / removed / changed facts), plus the full chronological supersession
-/// ledger (each: old object → new object, when it took effect, the source meeting). `entity` is a
-/// name OR an id — resolved through the GATED [`crate::summarize::dossier::resolve_entity_id`] (an
-/// entity mentioned only in sealed-not-unlocked meetings never resolves). The facts themselves are
-/// read through the visibility-gated [`crate::storage::Db::list_facts_visible`] inside
-/// [`crate::facts::build_knowledge_diff`] — a sealed-and-not-session-unlocked meeting's fact enters
-/// no snapshot, diff entry, or ledger row. Snapshots the live session unlock set like `get_graph`.
-/// Read-only, deterministic, no model, no new egress. `InvalidArg` if the entity does not resolve.
-#[tauri::command]
-pub fn get_entity_knowledge_diff(
-    state: State<'_, AppState>,
-    entity: String,
-    from: String,
-    to: String,
-) -> Result<crate::facts::EntityKnowledgeDiff, AppError> {
-    let unlocked = unlocked_snapshot(state.inner())?;
-    // GATE (anti-leak, FIRST): resolve name/id → a VISIBLE entity id, or refuse. An entity whose
-    // mentions are all sealed-not-unlocked is indistinguishable from an unknown one.
-    let id = crate::summarize::dossier::resolve_entity_id(&state.db, &entity, &unlocked)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no visible entity matching {entity:?}")))?;
-    crate::facts::build_knowledge_diff(&state.db, &id, &from, &to, &unlocked)
-}
-
-/// "What links here" for a note or meeting: every VISIBLE meeting-note / standalone-note whose body
-/// carries a `[[<this row's title>]]` wikilink. GATE is folded into the DB builder
-/// (`Db::backlinks_for_visible`) exactly like [`get_entity_detail`] — no separate command-layer
-/// pre-check. Snapshots the LIVE session unlock set, so a sealed target yields `[]` (never reveals it
-/// HAS backlinks) and a sealed source never contributes. `target_kind` is `"meeting"` | `"note"`.
-#[tauri::command]
-pub fn get_backlinks(
-    state: State<'_, AppState>,
-    target_kind: String,
-    target_id: String,
-) -> Result<Vec<crate::storage::models::BacklinkSource>, AppError> {
-    let kind = match target_kind.as_str() {
-        "meeting" => crate::storage::models::SourceKind::Meeting,
-        "note" => crate::storage::models::SourceKind::Note,
-        other => {
-            return Err(AppError::InvalidArg(format!(
-                "unknown backlink target_kind {other:?} (expected \"meeting\" or \"note\")"
-            )))
-        }
-    };
-    let unlocked = unlocked_snapshot(state.inner())?;
-    state.db.backlinks_for_visible(kind, &target_id, &unlocked)
 }
 
 /// Parse an IPC link-endpoint kind string into [`crate::links::LinkKind`], or a clean `InvalidArg`.
@@ -8414,22 +7475,6 @@ pub fn list_link_candidates(
         }
     }
     Ok(out)
-}
-
-/// Structured, GATED, egress-free person dossier for the `/people` detail pane. Unlike
-/// [`entity_dossier`] (which CLOUD-synthesizes a markdown String via the provider and discards the
-/// struct), this returns the STRUCTURED [`DossierData`](crate::summarize::dossier::DossierData) with
-/// NO provider/cloud call — deterministic DB assembly, strictly MORE local-first. Gated exactly like
-/// [`get_entity_detail`]/[`list_people`]: it snapshots the LIVE session unlock set and reuses
-/// `build_dossier_data` VERBATIM, so a sealed-and-not-session-unlocked meeting contributes NOTHING
-/// (its title, note body, commitments, and facts all stay invisible until the folder is
-/// session-unlocked). `corpus` is `#[serde(skip)]`, so meeting note bodies never reach the FE.
-#[tauri::command]
-pub fn get_person_dossier(
-    state: State<'_, AppState>,
-    entity_id: String,
-) -> Result<crate::summarize::dossier::DossierData, AppError> {
-    get_person_dossier_inner(state.inner(), &entity_id)
 }
 
 /// Inner of [`get_person_dossier`] taking `&AppState` (unit-testable gate). `None` — an unknown id
@@ -9616,102 +8661,6 @@ pub async fn entity_dossier(
     })
 }
 
-/// The testable, SYNCHRONOUS core of [`entity_dossier`]: resolve the entity + gated local dossier
-/// data, ALSO run the READ-ONLY org leg (B2), and assemble the (system, user, has_org_context)
-/// synthesis-prompt triple — everything short of the actual `provider.complete` cloud call, so this
-/// can be tested without a real provider/network.
-///
-/// B2 (Shared Brain, READ-ONLY): when the caller has joined an org (`org_brain_available`), ALSO
-/// searches the org partition ONCE for this entity and folds `[org · author]`-labelled hits into
-/// the synthesis prompt ONLY — an entity with ZERO local facts/mentions can still resolve here
-/// purely from org content (`data` may be `None` while `org_context` is `Some`). This function NEVER
-/// calls `build_and_persist_entities` and NEVER writes to `entities`/`entity_mentions`/`facts` — it
-/// only reads. `has_org_context` is `true` iff org-sourced lines were folded in, so the caller can
-/// honestly signal that the dossier drew on colleague content.
-fn build_entity_dossier_prompt(
-    db: &crate::storage::Db,
-    entity: &str,
-    unlocked: &std::collections::HashSet<String>,
-    config: &AppConfig,
-    notes_conn: &str,
-) -> Result<(String, String, bool), AppError> {
-    if entity.trim().is_empty() {
-        return Err(AppError::InvalidArg("entity is empty".into()));
-    }
-    // B2: resolve the ORG leg once, up front — an entity that is ORG-ONLY (no local entity row at
-    // all) must still be resolvable, so the local-entity lookup below is allowed to miss as long as
-    // org content is present. Gated on the SAME `org_brain_available` predicate as every other org
-    // leg (A1/A2); one call, no per-query looping. READ-ONLY: `search_org_brain` never writes.
-    let org_query = entity.trim().to_string();
-    let org_context = if crate::tools::org_brain_available(db, config) {
-        match crate::tools::search_org_brain(db, config, &org_query) {
-            Ok(text) => {
-                let text = text.trim();
-                // `search_org_brain` degrades to a "No org-brain results …" sentinel (never an Err)
-                // when there is nothing to find — that carries no `- [org` line, so it is dropped
-                // here exactly like every other "nothing found" sentinel this codebase filters
-                // (`voice_action::is_empty_tool_result`). A real hit's text is passed THROUGH
-                // WHOLE (never re-split by line) — a snippet can legitimately wrap onto a second
-                // line that doesn't itself start with "- [org", and truncating there would silently
-                // drop cited content from the synthesis prompt.
-                if text.is_empty() || !text.contains("- [org") {
-                    None
-                } else {
-                    Some(text.to_string())
-                }
-            }
-            Err(e) => {
-                tracing::debug!(
-                    target: "dossier",
-                    error = %e,
-                    "entity dossier org-brain search failed; continuing without org context"
-                );
-                None
-            }
-        }
-    } else {
-        None
-    };
-
-    let entity_id = crate::summarize::dossier::resolve_entity_id(db, entity, unlocked)?;
-    let data = match entity_id {
-        Some(id) => crate::summarize::dossier::build_dossier_data(db, &id, unlocked)?,
-        None => None,
-    };
-    // Neither a local entity NOR any org context ⇒ genuinely unknown — the pre-org error semantics.
-    if data.is_none() && org_context.is_none() {
-        return Err(AppError::InvalidArg(format!(
-            "no visible entity matching \"{entity}\""
-        )));
-    }
-    let mut system = crate::summarize::dossier::dossier_system_prompt(&config.note_language);
-    let mut user = match &data {
-        Some(d) => crate::summarize::dossier::render_dossier_user(d, notes_conn),
-        // Org-only entity: no local dossier data at all — hand the model just the entity name so it
-        // has a subject to synthesize the org section against.
-        None => format!(
-            "ENTITY: {}\n(no local meetings/facts found for this entity)\n",
-            entity.trim()
-        ),
-    };
-    let has_org_context = if let Some(org_lines) = &org_context {
-        system.push_str(
-            " Additional context may include lines starting \"- [org ·\" — these are colleague-\
-             shared facts from your organization's Shared Brain, NOT the user's own verified data. \
-             Attribute any claim drawn from them by their \"[org · author]\" provenance, clearly \
-             distinguished from the user's own meetings/facts.",
-        );
-        user.push_str(&format!(
-            "\n\nADDITIONAL CONTEXT FROM YOUR ORGANIZATION'S SHARED BRAIN (colleague-shared, \
-             attribute by provenance):\n{org_lines}\n"
-        ));
-        true
-    } else {
-        false
-    };
-    Ok((system, user, has_org_context))
-}
-
 #[cfg(test)]
 mod entity_dossier_tests {
     use super::*;
@@ -10241,79 +9190,6 @@ pub fn topic_threads(state: State<'_, AppState>) -> Result<Vec<TopicThread>, App
         });
     }
     Ok(crate::summarize::threads::build_threads(&input))
-}
-
-/// Export a meeting as an Obsidian Canvas (.canvas) — a spatial board of its topic spans.
-/// Requires the timeline (open the meeting once). Returns the written path.
-#[tauri::command]
-pub fn export_canvas(state: State<'_, AppState>, meeting_id: String) -> Result<String, AppError> {
-    // D4 READ-GATE: a sealed-and-not-unlocked meeting's timeline is blanked; refuse to build a
-    // canvas from it. Fail closed.
-    if !meeting_is_unlocked(state.inner(), &meeting_id)? {
-        return Err(AppError::Locked(
-            "this meeting's folder is locked — unlock it to export the canvas".into(),
-        ));
-    }
-    let meeting = state
-        .db
-        .get_meeting(&meeting_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no meeting with id {meeting_id}")))?;
-    let json = state.db.get_timeline_data(&meeting_id)?.ok_or_else(|| {
-        AppError::InvalidArg("open the meeting once to generate its timeline first".into())
-    })?;
-    let mut tl: MeetingTimeline = serde_json::from_str(&json)
-        .map_err(|e| AppError::InvalidArg(format!("bad timeline data: {e}")))?;
-    // Same coverage-repair the Detail view applies (heal a legacy cache that ends short of the
-    // recording) so the exported canvas spans the meeting instead of the provider's early cluster.
-    let segments = state.db.get_segments(&meeting_id)?;
-    crate::summarize::timeline::repair_coverage(&mut tl, &segments);
-    let title = meeting.title.unwrap_or_else(|| "Meeting".to_string());
-    let topics: Vec<(String, f64, f64)> = tl
-        .topics
-        .iter()
-        .map(|t| (t.label.clone(), t.start_s, t.end_s))
-        .collect();
-    let canvas = crate::export::canvas::build_canvas(&title, &topics);
-    let vault = {
-        state
-            .config
-            .lock()
-            .map_err(|_| AppError::Config("config mutex poisoned".into()))?
-            .vault_path
-            .clone()
-    }
-    .filter(|p| !p.is_empty())
-    .ok_or_else(|| AppError::InvalidArg("set a vault folder in Settings first".into()))?;
-    let vault_root = std::path::Path::new(&vault);
-    // D5: the Canvas dir must resolve inside the vault root.
-    let dir = assert_in_vault(vault_root, std::path::Path::new("Canvas"))?;
-    std::fs::create_dir_all(&dir)
-        .map_err(|e| AppError::Export(format!("create Canvas dir failed: {e}")))?;
-    let fname: String = title
-        .chars()
-        .map(|c| match c {
-            '/' | '\\' | ':' | '*' | '?' | '"' | '<' | '>' | '|' | '#' | '^' | '[' | ']' => ' ',
-            c if c.is_control() => ' ',
-            c => c,
-        })
-        .collect::<String>()
-        .split_whitespace()
-        .collect::<Vec<_>>()
-        .join(" ");
-    let fname = if fname.is_empty() {
-        "Meeting".to_string()
-    } else {
-        fname
-    };
-    // D5: re-assert the final file path stays inside the vault (fname is sanitized, but bind the
-    // guarantee at the write site).
-    let path = assert_in_vault(
-        vault_root,
-        &std::path::Path::new("Canvas").join(format!("{fname}.canvas")),
-    )?;
-    std::fs::write(&path, canvas)
-        .map_err(|e| AppError::Export(format!("write canvas failed: {e}")))?;
-    Ok(path.to_string_lossy().to_string())
 }
 
 /// Best-effort: the soonest macOS Calendar event in the next 60 minutes (title only). Returns
@@ -11145,127 +10021,6 @@ pub async fn gateway_health(state: State<'_, AppState>) -> Result<GatewayHealthD
 // All structs are `camelCase` on the wire (matches the FE `EgressLedger` / `EgressRow` types in
 // `core/models.ts`). Carries ONLY counts, ids, labels, and token/byte numbers — no content (§8).
 
-/// Per-model token-usage roll-up for `EgressLedger.byModel`.
-///
-/// Fields are `u64` so a large all-time cumulative sum (mirroring `EgressModelUsage`) cannot
-/// silently wrap. JavaScript `number` (f64) handles these values without precision loss up to
-/// 2^53 (~9 petaTokens), which is far beyond any realistic usage window.
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ModelUsageDto {
-    pub model: String,
-    pub calls: u64,
-    pub tokens: u64,
-}
-
-/// Per-day token-usage roll-up for `EgressLedger.byDay`.
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DayUsageDto {
-    /// ISO-8601 date string ("YYYY-MM-DD") in UTC.
-    pub day: String,
-    pub tokens: u64,
-}
-
-/// Redaction-count totals for the queried window.
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RedactionTotalsDto {
-    pub email: u64,
-    pub card: u64,
-    pub phone: u64,
-    pub name: u64,
-}
-
-/// One row from the `egress_log` table (content-free: counts + ids only).
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct EgressRowDto {
-    /// Unix epoch (seconds) of the call.
-    pub ts: i64,
-    pub provider_id: String,
-    pub destination: String,
-    pub model_served: Option<String>,
-    pub total_tokens: Option<u32>,
-    pub redactions: RedactionTotalsDto,
-}
-
-/// Aggregated egress ledger for a rolling window (`days` days back from now).
-///
-/// Shape matches `EgressLedger` in `src/app/core/models.ts` (camelCase).
-/// Every aggregate handles an empty `egress_log` gracefully — totals are zero, vecs empty.
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct EgressLedgerDto {
-    pub total_calls: u64,
-    pub total_tokens: u64,
-    pub by_model: Vec<ModelUsageDto>,
-    pub by_day: Vec<DayUsageDto>,
-    pub total_redactions: RedactionTotalsDto,
-    /// Last ≤20 rows from `egress_log`, newest first.
-    pub recent: Vec<EgressRowDto>,
-}
-
-/// Aggregate the content-free `egress_log` table for the given rolling window and return the
-/// ledger for the "Egress & Usage" Analytics panel.
-///
-/// `days` is the window width; pass `30` for the default 30-day view. The window is computed as
-/// `ts >= (now_unix - days * 86400)`. An empty table (no cloud calls yet) returns all-zero totals
-/// and empty vecs — never an error.
-///
-/// Read-only: queries `egress_log` only. No content columns are touched.
-#[tauri::command]
-pub fn get_egress_ledger(
-    days: i64,
-    state: State<'_, AppState>,
-) -> Result<EgressLedgerDto, AppError> {
-    let ledger = state.db.egress_summary(days)?;
-    Ok(EgressLedgerDto {
-        total_calls: ledger.total_calls,
-        total_tokens: ledger.total_tokens,
-        by_model: ledger
-            .by_model
-            .into_iter()
-            .map(|m| ModelUsageDto {
-                model: m.model,
-                calls: m.calls,
-                tokens: m.tokens,
-            })
-            .collect(),
-        by_day: ledger
-            .by_day
-            .into_iter()
-            .map(|d| DayUsageDto {
-                day: d.day,
-                tokens: d.tokens,
-            })
-            .collect(),
-        total_redactions: RedactionTotalsDto {
-            email: ledger.total_redactions.email,
-            card: ledger.total_redactions.card,
-            phone: ledger.total_redactions.phone,
-            name: ledger.total_redactions.name,
-        },
-        recent: ledger
-            .recent
-            .into_iter()
-            .map(|r| EgressRowDto {
-                ts: r.ts,
-                provider_id: r.provider_id,
-                destination: r.destination,
-                model_served: r.model_served,
-                total_tokens: r.total_tokens,
-                redactions: RedactionTotalsDto {
-                    email: r.redactions.email,
-                    card: r.redactions.card,
-                    phone: r.redactions.phone,
-                    name: r.redactions.name,
-                },
-            })
-            .collect(),
-    })
-}
-
 /// availability() fan-out across all three providers for the Settings UI.
 #[tauri::command]
 pub async fn provider_statuses(
@@ -11513,16 +10268,6 @@ fn mask_locked_meetings(
             }
         })
         .collect())
-}
-
-/// Aggregate analytics for the dashboard + Analytics tab. VISIBLE-content only — a sealed-and-
-/// not-session-unlocked folder's meetings are excluded from every count/duration/breakdown (same
-/// gate as `brain_overview`/`list_meetings`), so the Analytics tab can never reveal the size or
-/// activity pattern of content the user has deliberately locked.
-#[tauri::command]
-pub fn get_analytics(state: State<'_, AppState>) -> Result<Analytics, AppError> {
-    let unlocked = unlocked_snapshot(state.inner())?;
-    state.db.analytics(&unlocked)
 }
 
 /// Rename a speaker across a meeting's cached timeline (e.g. "User 1" → "Sarah"). Persists to
@@ -15742,7 +14487,7 @@ fn unseal_folder_extras_permanent(
 /// Snapshot the live session unlock set (the same source `list_folders` / the graph reads use).
 /// Passed to the `*_visible` DB reads (BLK-2b) so a sealed-and-not-unlocked meeting contributes
 /// nothing to digests, search, last-note, topic threads, etc. — independent of at-rest blanking.
-fn unlocked_snapshot(state: &AppState) -> Result<std::collections::HashSet<String>, AppError> {
+pub(crate) fn unlocked_snapshot(state: &AppState) -> Result<std::collections::HashSet<String>, AppError> {
     Ok(state
         .unlocked_folders
         .lock()
@@ -16037,7 +14782,7 @@ fn vault_path(state: &AppState) -> Option<String> {
 /// about to be created) — the deepest EXISTING ancestor is canonicalized (so symlinks are resolved)
 /// and the remaining components are appended after rejecting any `..` / root / prefix component that
 /// could climb out. The vault root itself must exist (it is the user-configured directory).
-fn assert_in_vault(
+pub(crate) fn assert_in_vault(
     vault: &std::path::Path,
     candidate: &std::path::Path,
 ) -> Result<std::path::PathBuf, AppError> {
